requirements.txt
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.1
idna==3.10
iniconfig==2.0.0
mysql-connector==2.2.9
packaging==24.2
paramiko==3.5.1
pluggy==1.5.0
pycparser==2.22
PyNaCl==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
sshtunnel==0.4.0
urllib3==2.3.0
